+++ bin/git-dev-insights.php
@@ -10,6 +10,7 @@ use GitDevInsights\FileAnalyzer\Plugins\Javascript\JsInlineScriptTagFileAnalyzer
 use GitDevInsights\FileAnalyzer\Plugins\PluginManager;
 use GitDevInsights\TrendGraph\BasicStatistics\FileExtensionChartHTMLFileGenerator;
 use GitDevInsights\TrendGraph\BasicStatistics\FocusChartHTMLFileGenerator;
+use GitDevInsights\TrendGraph\BasicStatistics\LanguageChartHTMLFileGenerator;
 
 require_once('vendor/autoload.php');
 
@@ -65,8 +66,8 @@ if (isset($argv[1]) && $argv[1] === '--config') {
     $fileExtensionChartGenerator = new FileExtensionChartHTMLFileGenerator($analysisResult, $projectConfigDataProvider);
     $fileExtensionChartGenerator->writeChartOutputToFile();
 
-    $languageChartGenerator = new GeneratorLanguageChartService($analysisResult, $projectConfigDataProvider);
-    $languageChartGenerator->generateOutputFile();
+    $languageChartGenerator = new LanguageChartHTMLFileGenerator($analysisResult, $projectConfigDataProvider);
+    $languageChartGenerator->writeChartOutputToFile();
 
     $focusChartGenerator = new GeneratorLanguageStackFocusChartService($analysisResult, $projectConfigDataProvider);
     $focusChartGenerator->generateOutputFile();
